add test and rework file structure
